Перенес всю логику из контроллера для задач в сервис

+++ backend/src/main/java/com/github/giga_chill/gigachill/service/TaskService.java
@@ -2,16 +2,16 @@ package com.github.giga_chill.gigachill.service;
 
 import com.github.giga_chill.gigachill.data.access.object.TaskDAO;
 import com.github.giga_chill.gigachill.data.transfer.object.TaskDTO;
+import com.github.giga_chill.gigachill.exception.BadRequestException;
 import com.github.giga_chill.gigachill.exception.ConflictException;
-import com.github.giga_chill.gigachill.exception.NotFoundException;
 import com.github.giga_chill.gigachill.mapper.TaskMapper;
 import com.github.giga_chill.gigachill.mapper.UserMapper;
 import com.github.giga_chill.gigachill.model.User;
+import com.github.giga_chill.gigachill.service.validator.*;
 import com.github.giga_chill.gigachill.util.UuidUtils;
 import com.github.giga_chill.gigachill.web.info.RequestTaskInfo;
 import com.github.giga_chill.gigachill.web.info.ResponseTaskInfo;
 import com.github.giga_chill.gigachill.web.info.ResponseTaskWithShoppingListsInfo;
-import jakarta.annotation.Nullable;
 import java.time.OffsetDateTime;
 import java.util.*;
 import lombok.RequiredArgsConstructor;
@@ -26,12 +26,20 @@ public class TaskService {
     private final TaskDAO taskDAO;
     private final ShoppingListsService shoppingListsService;
     private final UserService userService;
-    private final ParticipantsService participantsService;
+    private final ParticipantService participantsService;
     private final EventService eventService;
     private final TaskMapper taskMapper;
     private final UserMapper userMapper;
+    private final EventServiceValidator eventServiceValidator;
+    private final ParticipantsServiceValidator participantsServiceValidator;
+    private final ShoppingListsServiceValidator shoppingListsServiceValidator;
+    private final TaskServiceValidator taskServiceValidator;
+    private final UserServiceValidator userServiceValidator;
 
     public List<ResponseTaskInfo> getAllTasksFromEvent(UUID eventId, UUID userId) {
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+
         return taskDAO.getAllTasksFromEvent(eventId).stream()
                 .map(taskMapper::toResponseTaskInfo)
                 .peek(
@@ -45,6 +53,10 @@ public class TaskService {
     }
 
     public ResponseTaskWithShoppingListsInfo getTaskById(UUID taskId, UUID eventId, UUID userId) {
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+
         var task = taskMapper.toResponseTaskWithShoppingListsInfo(taskDAO.getTaskById(taskId));
         task.setPermissions(taskPermissions(eventId, taskId, userId));
         task.getShoppingLists()
@@ -65,30 +77,25 @@ public class TaskService {
     }
 
     public String createTask(UUID eventId, User user, RequestTaskInfo requestTaskInfo) {
+        var executorId =
+                requestTaskInfo.executorId() != null
+                        ? UuidUtils.safeUUID(requestTaskInfo.executorId())
+                        : null;
+
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        participantsServiceValidator.checkIsParticipant(eventId, user.getId());
+        if (executorId != null) {
+            userServiceValidator.checkIsExisted(executorId);
+            participantsServiceValidator.checkIsParticipant(eventId, executorId);
+        }
         var shoppingListsIds =
                 requestTaskInfo.shoppingListsIds().stream().map(UuidUtils::safeUUID).toList();
-
-        if (!shoppingListsService.areExisted(shoppingListsIds)) {
-            throw new NotFoundException(
-                    "One or more of the resources involved were not found: "
-                            + requestTaskInfo.shoppingListsIds());
-        }
-        if (!shoppingListsService.canBindShoppingListsToTask(shoppingListsIds)) {
-            throw new ConflictException(
-                    "One or more lists are already linked to the task: "
-                            + requestTaskInfo.shoppingListsIds());
-        }
-
+        shoppingListsServiceValidator.checkAreExisted(shoppingListsIds);
+        shoppingListsServiceValidator.checkOpportunityToBindShoppingListsToTask(shoppingListsIds);
         var eventEndDatetime = OffsetDateTime.parse(eventService.getEndDatetime(eventId));
         var taskDeadline = OffsetDateTime.parse(requestTaskInfo.deadlineDatetime());
-
-        if (eventEndDatetime.isBefore(taskDeadline)) {
-            throw new ConflictException(
-                    "You cannot specify task due date: "
-                            + taskDeadline
-                            + " that is later than the end of the event: "
-                            + eventEndDatetime);
-        }
+        taskServiceValidator.checkTaskDeadline(eventEndDatetime, taskDeadline);
 
         var task =
                 new TaskDTO(
@@ -105,12 +112,20 @@ public class TaskService {
                                         userService.getById(
                                                 UuidUtils.safeUUID(requestTaskInfo.executorId())))
                                 : null);
-
         taskDAO.createTask(eventId, task, shoppingListsIds);
         return task.getTaskId().toString();
     }
 
-    public void updateTask(UUID eventId, UUID taskId, RequestTaskInfo requestTaskInfo) {
+    public void updateTask(
+            UUID eventId, UUID taskId, UUID userId, RequestTaskInfo requestTaskInfo) {
+
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        taskServiceValidator.checkNotCompletedStatus(taskId, getTaskStatus(taskId));
+        participantsServiceValidator.checkIsAuthorOrAdminOrOwner(eventId, userId, taskId);
+
         var eventEndDatetime = OffsetDateTime.parse(eventService.getEndDatetime(eventId));
         var taskDeadline = OffsetDateTime.parse(requestTaskInfo.deadlineDatetime());
 
@@ -135,11 +150,25 @@ public class TaskService {
         taskDAO.updateTask(taskId, task);
     }
 
-    public void startExecuting(UUID taskId, UUID userId) {
+    public void startExecuting(UUID taskId, UUID userId, UUID eventId) {
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        taskServiceValidator.checkNotCompletedStatus(taskId, getTaskStatus(taskId));
+        taskServiceValidator.checkExecutionOpportunity(taskId, userId);
+
         taskDAO.startExecuting(taskId, userId);
     }
 
-    public void deleteTask(UUID taskId) {
+    public void deleteTask(UUID taskId, UUID eventId, UUID userId) {
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        taskServiceValidator.checkNotCompletedStatus(taskId, getTaskStatus(taskId));
+        participantsServiceValidator.checkIsAuthorOrAdminOrOwner(eventId, userId, taskId);
+
         taskDAO.deleteTask(taskId);
     }
 
@@ -163,19 +192,84 @@ public class TaskService {
         return taskDAO.getExecutorId(taskId);
     }
 
-    public void updateExecutor(UUID taskId, @Nullable UUID executorId) {
+    public UUID updateExecutor(UUID taskId, UUID eventId, UUID userId, Map<String, Object> body) {
+
+        if (!body.containsKey("executor_id")) {
+            throw new BadRequestException("Invalid request body: " + body);
+        }
+        UUID executorId =
+                body.get("executor_id") != null
+                        ? UuidUtils.safeUUID((String) body.get("executor_id"))
+                        : null;
+
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        participantsServiceValidator.checkIsAuthorOrAdminOrOwner(eventId, userId, taskId);
+        if (executorId != null) {
+            userServiceValidator.checkIsExisted(executorId);
+            participantsServiceValidator.checkIsParticipant(eventId, executorId);
+        }
+
         taskDAO.updateExecutor(taskId, executorId);
+        return executorId;
     }
 
-    public void updateShoppingLists(UUID taskId, List<UUID> shoppingLists) {
-        taskDAO.updateShoppingLists(taskId, shoppingLists);
+    public void updateShoppingLists(UUID taskId, UUID eventId, UUID userId, List<String> body) {
+        var shoppingListsIds =
+                body != null ? body.stream().map(UuidUtils::safeUUID).toList() : null;
+
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        taskServiceValidator.checkNotCompletedStatus(taskId, getTaskStatus(taskId));
+        participantsServiceValidator.checkIsAuthorOrAdminOrOwner(eventId, userId, taskId);
+        if (shoppingListsIds != null) {
+            shoppingListsServiceValidator.checkAreExisted(shoppingListsIds);
+            shoppingListsServiceValidator.checkOpportunityToBindShoppingListsToTask(
+                    shoppingListsIds);
+        }
+
+        taskDAO.updateShoppingLists(taskId, shoppingListsIds);
     }
 
-    public void setExecutorComment(UUID taskId, String executorComment) {
+    public String setExecutorComment(
+            UUID taskId, UUID eventId, UUID userId, Map<String, String> body) {
+
+        var executorComment = body.get("executor_comment");
+        if (executorComment == null) {
+            throw new BadRequestException("Invalid request body: " + body);
+        }
+
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        taskServiceValidator.checkInProgressStatus(taskId, getTaskStatus(taskId));
+        taskServiceValidator.checkOpportunityToSentTaskToReview(taskId, userId);
+
         taskDAO.setExecutorComment(taskId, executorComment);
+        return executorComment;
     }
 
-    public void setReviewerComment(UUID taskId, String reviewerComment, boolean isApproved) {
+    public void setReviewerComment(
+            UUID taskId, Map<String, Object> body, UUID eventId, UUID userId) {
+
+        var reviewerComment = (String) body.get("reviewer_comment");
+        var isApproved = (Boolean) body.get("is_approved");
+        if (reviewerComment == null || isApproved == null) {
+            throw new BadRequestException("Invalid request body: " + body);
+        }
+
+        eventServiceValidator.checkIsExistedAndNotDeleted(eventId);
+        eventServiceValidator.checkIsFinalized(eventId);
+        taskServiceValidator.checkIsExisted(eventId, taskId);
+        participantsServiceValidator.checkIsParticipant(eventId, userId);
+        taskServiceValidator.checkUnderReviewStatus(taskId, getTaskStatus(taskId));
+        taskServiceValidator.checkOpportunityToApproveTask(eventId, taskId, userId);
+
         taskDAO.setReviewerComment(taskId, reviewerComment, isApproved);
     }
